requirements.txt
google-genai
pymupdf
tqdm